Set version dependent attributes in the separated recipe
Attributes, which default value depends on the value of another attribute, should be defined in the recipe, not in attribute file.
Otherwise, it is very hard to override such attributes or wrap this cookbook in another one.

Example:
Attributes `checksum` and `url` depend on `version` value. If you want to override `version`, then you should override
`checksum` and `url` as well, because their default values are calculated on the step of attribute file parsing, e.q.
before `version` value will be explored according to the priority.

This is so called "Derived Attributes" issue, read more: https://coderanger.net/derived-attributes/