Disable https on the Tomcat side
Production instances of Confluence usually use some web server as a reverse proxy at the front of Tomcat.
In this cookbook we also configure Apache web server for these needs.

So, since https is actually terminating on the Apache side, there are no needs to manage keystore
and configure TLS/SSL on the Tomcat side.