Update setenv.sh.erb update JRE_HOME condition
By updating the AND condition to "bundled_jre = true" (which is the default value), it allows the condition to evaluate to false by setting attribute "bundled_jre = false". This allows the java cookbook's java_home attribute to be set, if it has been overridden. 

Without the "= true" condition, the IF condition can't be bypassed to the ELSE condition, if desired.